Split digital.py to clock.py and digital_display.py and Updated Clock() Clock() has been updated to new Connector() class Clock.connect() and Clock.disconnect() have been added
