+++ BinPy/tools/clock.py
@@ -1,7 +1,7 @@
 import sys
 import time
 import threading
-from BinPy import Connector
+from BinPy.Gates import *
 
 class Clock(threading.Thread):
     """
@@ -26,9 +26,9 @@ class Clock(threading.Thread):
         init_state:     It is the initial state of the clock(1 by default)
         name:           It is the name of the clock.(optional)
     
-    Methods :   start(), getState(), setState(value), getName(), getTimePeriod(), kill()
-
-    """
+    Methods :   start(), connect(*connectors), disconnect(*connectors),
+                getState(), setState(value), getName(), getTimePeriod(), kill()    """
+    
     def __init__(self, init_state=1, frequency=None, time_period=None, name=None):
         threading.Thread.__init__(self)
         if frequency != None:
@@ -42,19 +42,45 @@ class Clock(threading.Thread):
         self.name = name
         self.curr_state = init_state
         self.exitFlag = 0
+        self.taps = []
         self.daemon = True
-        self.A = Connector(self, 'A', activates=1, monitor=0)
+
+    def connect(self, *connectors):
+
+        for connector in connectors:
+            if not isinstance(connector, Connector):
+                raise Exception("Error: Input given is not a connector")
+            else:
+                if len(connector.connections['output']) != 0:
+                    raise Exception("ERROR: The connector is already an output of some other object")
+                self.taps.append(connector)
+                connector.state = self.curr_state
+                connector.tap(self, 'output')
+                connector.trigger()              
+
+    def disconnect(self, *connectors):
+
+        for connector in connectors:
+            if isinstance(connector, Connector):
+                try:
+                    self.taps.remove(Connector)
+                    connector.state = None
+                    connector.connections['output'].remove(self)
+                    connector.trigger()
+                except:
+                    print ("The specified connector is not taped to this clock")
+            else:
+                raise Exception("Error: Input given is not a connector")
+
 
     def __toggleState(self):
         """
         This is an internal method to toggle the state of the output
         """
-        if self.curr_state==1:
-            self.curr_state = 0
-            self.A.set(self.curr_state)
-        else:
-            self.curr_state = 1
-            self.A.set(self.curr_state)
+        self.curr_state = not self.curr_state
+        for connector in self.taps:
+            connector.trigger()
+        
 
     def __main_func(self):
         while True:
@@ -99,68 +125,3 @@ class Clock(threading.Thread):
 
     def run(self):
         self.__main_func()
-
-
-class DigitDisplay:
-    '''
-    This class emulates a 7 segmented display(Common Cathode)
-
-    Parameters:
-        name:   A name given to an object(Optional)
-
-    Methods:
-        evaluate()
-        getName()
-
-    How to use:
-        >>> myDisplay = DigitDisplay("Display1")
-        >>> print myDisplay.evaluate([1,1,1,1,1,1,1])
-        8
-    Note:
-        You can either pass complete list of 10 pins [pin1, pin2, pin3, pin4, pin5, pin6, pin7, pin8, pin9, pin10]
-        in standard order (see http://tronixstuff.files.wordpress.com/2010/05/7segpinout.jpg) or you can directly
-        pass the list of values corresponding to a, b, c, d, e, f and g in lexicographical order.
-    '''
-    def __init__(self,name=None):
-        self.name = name
-
-    def evaluate(self,pin_conf):
-        '''
-        This method evaluates the values passed according to the display and returns
-        an integer varying from 0 to 9
-        '''
-        if len(pin_conf)!=10:
-            if len(pin_conf)!=7:
-                raise Exception("There must be 10 or 7 values")
-        if len(pin_conf)==10:
-            vcc = pin_conf[2] or pin_conf[7]
-            a = pin_conf[6]
-            b = pin_conf[5]
-            c = pin_conf[3]
-            d = pin_conf[1]
-            e = pin_conf[0]
-            f = pin_conf[8]
-            g = pin_conf[9]
-        if len(pin_conf)==7:
-            a = pin_conf[0]
-            b = pin_conf[1]
-            c = pin_conf[2]
-            d = pin_conf[3]
-            e = pin_conf[4]
-            f = pin_conf[5]
-            g = pin_conf[6]
-            vcc=1           
-        if vcc:
-            test = [a, b, c, d, e, f, g]
-            data = {'0':[1,1,1,1,1,1,0],'1':[0,1,0,0,0,0,0],'2':[1,1,0,1,1,0,1],\
-            '3':[1,1,1,1,0,0,1],'4':[0,1,1,0,0,1,1],'5':[1,0,1,1,0,1,1],'6':[1,0,1,1,1,1,1],\
-            '7':[1,1,1,0,0,0,0],'8':[1,1,1,1,1,1,1],'9':[1,1,1,1,0,1,1]}
-            for i in data:
-                if test==data[i]:
-                    return int(i)
-            return None
-        else:
-            return None
-
-        def getName(self):
-            return self.name
